Run eslint at 'prepublish' time
We do not want to run eslint whenever this package is installed, but rather we want to ensure that versions
we publish are proper. This avoids requiring to install the devDependencies in all cases, i.e. makes `npm install --production`/`NODE_ENV=production npm install`
possible again.

Note that this requires that the cleanup step in Travis is skipped for deploys, which as a side-effect
also improves the build time when deploys are done.